chore: refactor backendAdapter and methods requiring backendAdapter to server-only.
